fixed pr comments
